Extracted Parsing logic into helper class
